Renamed some files for convention
